Cleanup around slash subcommands
